CMAKE status message ++.

+++ unittests/type_limits/test_type_limits.cu
@@ -4,7 +4,7 @@
 #include <gunrock/util/type_limits.hxx>
 
 void test_type_limits() {
-  using type_t = short unsigned int;
+  using type_t = unsigned int;
   type_t i = gunrock::numeric_limits<type_t>::invalid();
   std::cout << "i = " << i << " (is valid? " << std::boolalpha
             << gunrock::util::limits::is_valid(i) << ")" << std::endl;
